% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object.R
\name{andromeda}
\alias{andromeda}
\title{Create an Andromeda object}
\usage{
andromeda()
}
\description{
Create an Andromeda object
}
\details{
By default the Andromeda object is created in the systems temporary file location. You can override this by 
specifying a folder using \code{options(andromedaTempFolder = "c:/andromedaTemp")}, where "c:/andromedaTemp"
is the folder to create the Andromeda objects in.
}
