% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operations.R
\name{appendToTable}
\alias{appendToTable}
\title{Append to an Andromeda table}
\usage{
appendToTable(tbl, data)
}
\arguments{
\item{tbl}{An Andromeda table. This must be a base table (i.e. it cannot be a query result).}

\item{data}{The data to append. This can be either a data.frame or another Andromeda table.}
}
\description{
Append a data frame, Andromeda table, or result of a query on an Andromeda table to an existing
Andromeda table.
If data from another Andromeda is appended, a batch-wise copy process is used, which will be slower
than when appending data from within the same Andromeda object.
}
\examples{
andr <- andromeda(cars = cars)
nrow(andr$cars)
# [1] 50

appendToTable(andr$cars, cars)
nrow(andr$cars)
# [1] 100

appendToTable(andr$cars, andr$cars \%>\% filter(speed > 10))
nrow(andr$cars)
# [1] 182

close(andr)

}
