% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operations.R
\name{batchApply}
\alias{batchApply}
\title{Apply a function to batches of data in an Andromeda table}
\usage{
batchApply(tbl, fun, ..., batchSize = 10000)
}
\arguments{
\item{tbl}{An Andromeda table.}

\item{fun}{A function where the first argument is a data frame.}

\item{batchSize}{Number of rows to fetch at a time.}
}
\description{
Apply a function to batches of data in an Andromeda table
}
\details{
This function is similar to the \code{lapply} function, in that it applies a 
function to sets of data. In this case, the data is batches of data from an
Andromeda table. Each batch will be presented to the function as a data frame.
}
