% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object.R
\name{$.Andromeda}
\alias{$.Andromeda}
\title{Extract table reference from Andromeda}
\usage{
\method{$}{Andromeda}(x, name)
}
\arguments{
\item{x}{An andromeda object}

\item{name}{The name of a table in the andromeda object}
}
\value{
An andromeda table
}
\description{
Extract table reference from Andromeda
}
