% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object.R
\name{$,Andromeda-method}
\alias{$,Andromeda-method}
\title{Extract table reference from Andromeda}
\usage{
\S4method{$}{Andromeda}(x, name)
}
\arguments{
\item{x}{An \code{\link{Andromeda}} object}

\item{name}{The name of a table in the andromeda object}
}
\value{
An andromeda table
}
\description{
Extract table reference from Andromeda
}
