% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadingSaving.R
\name{checkAndromedaTempFolder}
\alias{checkAndromedaTempFolder}
\title{Check that Andromeda temp folder}
\usage{
checkAndromedaTempFolder(minimumSize)
}
\arguments{
\item{minimumSize}{(Optional) The minimum required number of available gigabytes for the Andromeda temp folder.}
}
\value{
This function is not called for its return value.
}
\description{
Checks that the Andromeda temp folder exists, is writable, and optionally has a minimum required amount of available disk space.
The Andromeda temp folder is the location where Andromeda objects are stored.
}
\examples{
# Check that the Andromeda temp folder has at least 10GB. 
checkTempFolder(minimumSize = 10)
}
