% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operations.R
\name{collectBatched}
\alias{collectBatched}
\title{Collect in batches}
\usage{
collectBatched(tbl, fun, batchSize = 10000)
}
\arguments{
\item{tbl}{An Andromeda table.}

\item{fun}{A function that takes a single data frame as argument.}

\item{batchSize}{Number of rows to fetch at a time.}
}
\description{
Collect in batches
}
