% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operations.R
\name{flushAndromeda}
\alias{flushAndromeda}
\title{Flush changes to disk}
\usage{
flushAndromeda(andromeda)
}
\arguments{
\item{andromeda}{An \code{\link{Andromeda}} object.}
}
\value{
This function does not return anything, but is executed for its side-effect.
}
\description{
Flush all changes to disk. This only affects the data in the Andromeda temp folder.

This function is for advanced users, who wish to share the same underlying object with other
processes, which is very dangerous if you don't know what you're doing.
}
\examples{
andr <- andromeda()
andr$cars <- cars

flushAndromeda(andr)
# The database in andromeda temp is now fully up-to-date.

close(andr)

}
