% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object.R
\name{isAndomeda}
\alias{isAndomeda}
\title{Check whether an object is an Andromeda object}
\usage{
isAndomeda(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
A logical value.
}
\description{
Check whether an object is an Andromeda object
}
\details{
Checks whether an object is an Andromeda object.
}
