% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object.R
\name{isAndromedaTable}
\alias{isAndromedaTable}
\title{Is the object an Andromeda table?}
\usage{
isAndromedaTable(tbl)
}
\arguments{
\item{tbl}{A reference to an Andromeda table}
}
\value{
TRUE or FALSE
}
\description{
Is the object an Andromeda table?
}
\examples{
andr <- andromeda(cars = cars)
isAndromedaTable(andr$cars)
close(andr)
}
