% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object.R
\name{isValidAndromeda}
\alias{isValidAndromeda}
\title{Checks whether an Andromeda object is valid}
\usage{
isValidAndromeda(x)
}
\arguments{
\item{x}{an Andromeda object}
}
\value{
TRUE if the Andromeda object is in a valid state. FALSE otherwise.
}
\description{
Checks whether an Andromeda object is valid
}
