% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object.R
\name{length,Andromeda-method}
\alias{length,Andromeda-method}
\title{Number of tables in an Andromeda object}
\usage{
\S4method{length}{Andromeda}(x)
}
\arguments{
\item{x}{An \code{\link{Andromeda}} object}
}
\value{
The number of tables in the andromeda object
}
\description{
Number of tables in an Andromeda object
}
