% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object.R
\name{length.Andromeda}
\alias{length.Andromeda}
\title{Number of tables in an Andromeda object}
\usage{
\method{length}{Andromeda}(x)
}
\arguments{
\item{x}{An andromeda object}
}
\value{
The number of tables in the andromeda object
}
\description{
Number of tables in an Andromeda object
}
