% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadingSaving.R
\name{loadAndromeda}
\alias{loadAndromeda}
\title{Load Andromeda from file}
\usage{
loadAndromeda(fileName)
}
\arguments{
\item{fileName}{The path where the object was saved using \code{\link{saveAndromeda}}.}
}
\description{
Load Andromeda from file
}
\examples{
\dontrun{
andr <- loadAndromeda("c:/temp/andromeda.zip")
names(andr)
# [1] "cars"

close(andr) 
}

}
\seealso{
\code{\link{saveAndromeda}}
}
