% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object.R
\name{names<-.tbl_Andromeda}
\alias{names<-.tbl_Andromeda}
\title{Set column names of an Andromeda table}
\usage{
\method{names}{tbl_Andromeda}(x) <- value
}
\arguments{
\item{x}{A reference to a table in an andromeda object. (see examples)}

\item{value}{A character vector of new names that must have length equat to the number of columns in the table.}
}
\description{
Set column names of an Andromeda table
}
\examples{
andr <- andromeda(cars = cars)
names(andr$cars) <- toupper(names(andr$cars))
names(andr$cars)
[1] "SPEED" "DIST" 
close(andr)
}
