% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object.R
\name{names.tbl_Andromeda}
\alias{names.tbl_Andromeda}
\title{Get the column names of an Andromeda table}
\usage{
\method{names}{tbl_Andromeda}(tbl)
}
\arguments{
\item{tbl}{An table in an Andromeda object}
}
\value{
A character vector of column names
}
\description{
Get the column names of an Andromeda table
}
\examples{
andr <- andromeda(cars = cars)
names(andr$cars)
[1] "speed" "dist"

}
