% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadingSaving.R
\name{saveAndromeda}
\alias{saveAndromeda}
\title{Save Andromeda to file}
\usage{
saveAndromeda(
  andromeda,
  fileName,
  maintainConnection = FALSE,
  overwrite = TRUE
)
}
\arguments{
\item{andromeda}{An object of class Andromeda.}

\item{fileName}{The path where the object will be written.}

\item{maintainConnection}{Should the connection be maintained after saving? If FALSE, the Andromeda object will be invalid after this operation, but
saving will be faster.}

\item{overwrite}{If the file exists, should it be overwritten? If FALSE and the file exists, an error will be thrown.}
}
\description{
Save Andromeda to file
}
