% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object.R
\name{[[.Andromeda}
\alias{[[.Andromeda}
\title{Extract Andromeda table}
\usage{
\method{[[}{Andromeda}(x, name)
}
\arguments{
\item{x}{An andromeda object}

\item{name}{A character string containing the name of an Andromeda table}
}
\value{

}
\description{
Extract Andromeda table
}
