% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object.R
\name{[[,Andromeda-method}
\alias{[[,Andromeda-method}
\title{Extract Andromeda table}
\usage{
\S4method{[[}{Andromeda}(x, name)
}
\arguments{
\item{x}{An \code{\link{Andromeda}} object}

\item{name}{A character string containing the name of an Andromeda table}
}
\value{
An \code{\link{Andromeda}} table
}
\description{
Extract Andromeda table
}
