% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object.R
\docType{class}
\name{Andromeda-class}
\alias{Andromeda-class}
\alias{show,Andromeda-method}
\title{Andromeda class.}
\usage{
\S4method{show}{Andromeda}(object)
}
\description{
Andromeda class.
}
\keyword{internal}
