% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object.R
\docType{class}
\name{Andromeda-class}
\alias{Andromeda-class}
\alias{Andromeda}
\alias{show,Andromeda-method}
\alias{$<-,Andromeda-method}
\alias{[[<-,Andromeda-method}
\alias{names,Andromeda-method}
\alias{close,Andromeda-method}
\title{The Andromeda class}
\usage{
\S4method{show}{Andromeda}(object)

\S4method{$}{Andromeda}(x, name) <- value

\S4method{[[}{Andromeda}(x, i) <- value

\S4method{names}{Andromeda}(x)

\S4method{close}{Andromeda}(con, ..., verbose = TRUE)
}
\arguments{
\item{object}{An \code{\link{Andromeda}} object.}

\item{x}{An andromeda object}

\item{name}{The name of a table in the \code{\link{Andromeda}} object.}

\item{value}{A dataframe, \code{\link{Andromeda}} table, or dplyr query that uses an \code{\link{Andromeda}} table.}

\item{i}{The name of a table in the \code{\link{Andromeda}} object.}

\item{con}{An \code{\link{Andromeda}} object.}

\item{...}{Included for compatibility with generic \code{close()} method.}

\item{verbose}{Should a message be printed if the attempt to remove the andromeda directory was unsuccessful (TRUE or FALSE). .}
}
\value{
A character vector with table names

0 for success, 1 for failure, invisibly. If the andromeda object is already closed (file does not exist) 0 is returned.
}
\description{
The \code{Andromeda} class is an S3 object.

This class is provides the ability to work with data objects in R that are too large to fit in memory. Instead,
these objects are stored on disk. This is slower than working from memory, but may be the only viable option.
It is essentially a list of arrow datasets that are stored on disk in the feather format.

Attempts to delete an andromeda object.
}
\section{Tables}{

An \code{Andromeda} object has zero, one or more tables. The list of table names can be retrieved using the \code{\link[=names]{names()}}
method. Tables can be accessed using the dollar sign syntax, e.g. \code{andromeda$myTable}, or double-square-bracket
syntax, e.g. \code{andromeda[["myTable"]]}
}

\section{Permanence}{


To mimic the behavior of in-memory objects, when working with data in \code{Andromeda} the data is stored in a
temporary location on the disk. You can modify the data as you can see fit, and when needed can save the data
to a permanent location. Later this data can be loaded to a temporary location again and be read and modified,
while keeping the saved data as is.
}

\seealso{
\code{\link[=andromeda]{andromeda()}}
}
