% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Andromeda.R
\docType{package}
\name{Andromeda-package}
\alias{Andromeda-package}
\title{Andromeda: AsynchroNous Disk-based Representation of MassivE DAta}
\description{
A replacement for ff for storing large data objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OHDSI/Andromeda}
  \item Report bugs at \url{https://github.com/OHDSI/Andromeda/issues}
}

}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org}

}
\keyword{internal}
