% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Andromeda.R
\docType{package}
\name{Andromeda-package}
\alias{Andromeda-package}
\title{Andromeda: Asynchronous Disk-Based Representation of Massive Data}
\description{
Storing very large data objects on a local drive, while still making it possible to manipulate the data in an efficient manner.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OHDSI/Andromeda}
  \item Report bugs at \url{https://github.com/OHDSI/Andromeda/issues}
}

}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org}

Authors:
\itemize{
  \item Adam Black \email{black@ohdsi.org}
  \item Egill Fridgeirsson \email{e.fridgeirsson@erasmusmc.nl}
  \item Marc A. Suchard \email{msuchard@ucla.edu}
}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
