% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object.R
\name{Andromeda-class}
\alias{Andromeda-class}
\alias{andromeda}
\alias{Andromeda}
\title{The Andromeda class}
\usage{
andromeda(...)
}
\arguments{
\item{...}{Named objects. See details for what objects are valid. If no objects are provided, an
empty Andromeda is returned.}
}
\value{
Returns an \code{\link{Andromeda}} object.
}
\description{
The \code{Andromeda} class is an S3 object.

This class provides the ability to work with data objects in R that are too large to fit in memory. Instead,
these objects are stored on disk. This is slower than working from memory, but may be the only viable option.
The Andromeda object is a thin wrapper around a dm object from the
\href{https://krlmlr.github.io/dm/articles/dm.html}{(data models) package} but is restricted to the duckdb database
backend.

By default the \code{Andromeda} object is created in the systems temporary file location. You can override
this by specifying a folder using \code{options(andromedaTempFolder = "c:/andromedaTemp")}, where
\code{"c:/andromedaTemp"} is the folder to create the Andromeda objects in.
}
\details{
Valid objects are data frames, \code{Andromeda} tables, or any other \code{\link{dplyr}} table.
}
\section{Tables}{

An \code{Andromeda} object can have zero, one, or more tables. The list of table names can be retrieved using the \code{\link[=names]{names()}}
method. Tables can be accessed using the dollar sign syntax, e.g. \code{andromeda$myTable}, or double-square-bracket
syntax, e.g. \code{andromeda[["myTable"]]}
}

\section{Permanence}{


To mimic the behavior of in-memory objects, when working with data in \code{Andromeda} the data is stored in a
temporary location on the disk. You can modify the data as you can see fit, and when needed can save the data
to a permanent location. Later this data can be loaded to a temporary location again and be read and modified,
while keeping the saved data as is.
}

\section{Inheritance}{


The \code{Andromeda} inherits directly from \code{dm} As such, it can be used as if it is a \code{SQLiteConnection}.
\code{\link{duckdb}} is an R wrapper around 'duckdb', a low-weight but very powerful single-user SQL database similar to
SQLite that can run from a single file on the local file system.
}

\examples{
andr <- andromeda(cars = cars, iris = iris)

names(andr)
# [1] 'cars' 'iris'

andr$cars \%>\% filter(speed > 10) \%>\% collect()
# # A tibble: 41 x 2 
# speed dist 
# <dbl> <dbl> 
# 1 11 17 
# ...

close(andr)

}
\seealso{
\code{\link[=andromeda]{andromeda()}}
Create an Andromeda object
}
