% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operations.R
\name{appendToTable}
\alias{appendToTable}
\title{Append to an Andromeda table}
\usage{
appendToTable(tbl, data)
}
\arguments{
\item{tbl}{An Andromeda table. This must be a base table (i.e. it cannot be a query result).}

\item{data}{The data to append. This can be either a data.frame or another Andromeda table.}
}
\description{
Append to an Andromeda table
}
