% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object.R
\name{copyAndromeda}
\alias{copyAndromeda}
\title{Copy Andromeda}
\usage{
copyAndromeda(
  andromeda,
  dbname = tempfile(tmpdir = .getAndromedaTempFolder(), fileext = ".duckdb")
)
}
\arguments{
\item{andromeda}{The \code{\link{Andromeda}} object to copy.}

\item{dbname}{The location where the new Andromeda object will be stored on disc. Defaults to temp directory.}
}
\value{
The copied \code{\link{Andromeda}} object.
}
\description{
Creates a complete copy of an \code{\link{Andromeda}} object. Object attributes are not copied.
}
\examples{
andr <- andromeda(cars = cars, iris = iris)

andr2 <- copyAndromeda(andr)

names(andr2)
# [1] 'cars' 'iris'

close(andr)
close(andr2)

}
