% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object.R
\name{isValidAndromeda}
\alias{isValidAndromeda}
\title{Check whether an Andromda object is still valid}
\usage{
isValidAndromeda(x)
}
\arguments{
\item{x}{The Andromeda object to check.}
}
\value{
A logical value.
}
\description{
Check whether an Andromda object is still valid
}
\details{
Checks whether an Andromeda object is still valid, or whether it has been closed.
}
\examples{
andr <- andromeda(cars = cars, iris = iris)

isValidAndromeda(andr)
# TRUE

close(andr)

isValidAndromeda(andr)
# FALSE

}
