% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadingSaving.R
\name{loadAndromeda}
\alias{loadAndromeda}
\title{Load Andromeda from file}
\usage{
loadAndromeda(fileName)
}
\arguments{
\item{fileName}{The path where the object is stored.}
}
\description{
Load Andromeda from file
}
