% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadingSaving.R
\name{loadAndromeda}
\alias{loadAndromeda}
\title{Load Andromeda from file}
\usage{
loadAndromeda(fileName)
}
\arguments{
\item{fileName}{The path where the object was saved using \code{\link[=saveAndromeda]{saveAndromeda()}}.}
}
\value{
An \code{\link{Andromeda}} object.
}
\description{
Load Andromeda from file
}
\examples{
\dontrun{
# For this example we create an Andromeda object and save it to
# a temporary file location
fileName <- tempfile()
andr <- andromeda(cars = cars)
saveAndromeda(andr, fileName)

# Using loadAndromeda to load the object back:
andr <- loadAndromeda(fileName)


# Cleaning up the file used in this example:
unlink(fileName)
}
}
\seealso{
\code{\link[=saveAndromeda]{saveAndromeda()}}
}
