% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object.R
\name{names,Andromeda-method}
\alias{names,Andromeda-method}
\title{names}
\usage{
\S4method{names}{Andromeda}(x)
}
\arguments{
\item{x}{An Andromeda object.}
}
\description{
Show the names of the tables in an Andromeda object.
}
