% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadingSaving.R
\name{saveAndromeda}
\alias{saveAndromeda}
\title{Save Andromeda to file}
\usage{
saveAndromeda(andromeda, fileName, overwrite = TRUE)
}
\arguments{
\item{andromeda}{An object of class \code{\link{Andromeda}}.}

\item{fileName}{The path where the object will be written.}

\item{overwrite}{If the file exists, should it be overwritten? If \code{FALSE} and the file
exists, an error will be thrown.}
}
\value{
Invisibly returns the andromeda object being saved.
}
\description{
Saves the \code{\link{Andromeda}} object in a zipped file.
Each dataset is represented as a folder contiaining one or more feather files.
User defined attributes are saved to a json file.
}
\examples{
andr <- andromeda(cars = cars)

# For this example we'll use a temporary file location:
fileName <- tempfile()

saveAndromeda(andr, fileName)

# Cleaning up the file used in this example:
unlink(fileName)

}
\seealso{
\code{\link{loadAndromeda}}

\code{\link[=loadAndromeda]{loadAndromeda()}}
}
