% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadingSaving.R
\name{saveAndromeda}
\alias{saveAndromeda}
\title{Save Andromeda to file}
\usage{
saveAndromeda(
  andromeda,
  fileName,
  maintainConnection = FALSE,
  overwrite = TRUE
)
}
\arguments{
\item{andromeda}{An object of class Andromeda.}

\item{fileName}{The path where the object will be written.}

\item{maintainConnection}{Should the connection be maintained after saving? If FALSE, the Andromeda object will be invalid after this operation, but
saving will be faster.}

\item{overwrite}{If the file exists, should it be overwritten? If FALSE and the file exists, an error will be thrown.}
}
\description{
Saves the Andromeda object in a zipped file. Note that by default the Andromeda object is automatically closed by saving it to disk. This is due
to a limitation of the underlying technology (RSQLite). To keep the connection open, use \code{maintainConnection = TRUE}. This will first create
a temporary copy of the Andromeda object. Note that this can be substantially slower.
}
\examples{
\dontrun{
andr <- andromeda(cars = cars)
saveAndromeda(cars, "c:/temp/andromeda.zip")
}

}
\seealso{
\code{\link{loadAndromeda}}
}
